% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterWidget.R
\name{rasterWidget}
\alias{rasterWidget}
\title{View raster objects interactively}
\usage{
rasterWidget(raster, width = 400, height = 400, nclass, style = "fisher",
  colors, colorsToInterpolate = c("yellow", "red"), colNA = "gray")
}
\arguments{
\item{raster}{a \code{Raster*} object.}

\item{width}{width of the widget in px}

\item{height}{height of the widget in px}

\item{nclass}{number of classes required}

\item{style}{chosen style: one of "fixed", "sd", "equal", "pretty", "quantile", "kmeans", "hclust", "bclust", "fisher", or "jenks"}

\item{colors}{a character vector for color coding the class intervals}

\item{colNA}{character for the color of NA values}

\item{colorToInterpolate}{colors to interpolate}
}
\description{
this function produces an interactive view of the specified
raster object to inspect its cell values when hovering over them.
}
\examples{
require(raster)
meuseFile <- system.file("external/test.grd", package="raster")
meuseRaster <- raster(meuseFile)
rasterWidget(meuseRaster)
rasterWidget(meuseRaster, 500, 550)
rasterWidget(meuseRaster, nclass=5)
rasterWidget(meuseRaster, nclass=5, colors=c('yellow', 'orange', 'darkorange', 'red', 'darkred'))
rasterWidget(meuseRaster, colorsToInterpolate=c("green", "red"))
install.packages("RColorBrewer")
require(RColorBrewer)
rasterWidget(meuseRaster, colorsToInterpolate=brewer.pal(7, "Paired"))

}

