% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterWidget.R
\name{rasterWidgetOutput}
\alias{rasterWidgetOutput}
\title{Widget output function for use in Shiny}
\usage{
rasterWidgetOutput(outputId, width = "100\%", height = "400px")
}
\description{
Widget output function for use in Shiny
}

