% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterWidget.R
\name{renderRasterWidget}
\alias{renderRasterWidget}
\title{Widget render function for use in Shiny}
\usage{
renderRasterWidget(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{expr}

\item{env}{env}

\item{quoted}{quoted}
}
\description{
Widget render function for use in Shiny
}

